+++ packages/application/src/use-cases/index.ts
@@ -1,2 +1,7 @@
 export * from './detect-workspace.js';
 export * from './run-pipeline.js';
+export * from './resume-run.js';
+export * from './stop-run.js';
+export * from './get-run-history.js';
+export * from './get-run-details.js';
+export * from './list-artifacts.js';